Work has begun on updating sampling and processing. Added in linear map to rescale the values for both the texture and the surface. Have just realised that the sphere uncertainty has slightly broken - it doesn't look as uniform as it should. Will investigate after lunch. First thing to do will be to check that registration isn't screwing things over. Make registration optional, add it to the pre-processing list along with inversion and erosion.

+++ Sams_Project/Plugins/com.sam.finalyearproject/src/internal/Sams_View.h
@@ -123,7 +123,7 @@ class Sams_View : public QmitkAbstractView {
 
     //  b
     mitk::Image::Pointer GenerateUncertaintyTexture();
-    int SampleUncertainty(vtkVector<float, 3> startPoint, vtkVector<float, 3> direction);
+    double SampleUncertainty(vtkVector<float, 3> startPoint, vtkVector<float, 3> direction);
 
     //  c
